# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bildmind Public API",
    "description": "API для відкритого модуля навчання платформи Bildmind.",
    "license": {
      "name": "CC BY-NC-ND 4.0",
      "url": "https://creativecommons.org/licenses/by-nc-nd/4.0/"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Локальний сервер розробки"
    }
  ],
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Отримати всіх користувачів",
        "description": "Повертає всіх активних користувачів (тільки адміну).",
        "operationId": "list_users_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserRead"
                  },
                  "type": "array",
                  "title": "Response List Users Users Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Отримати користувача за id",
        "description": "Повертає користувача по id (себе або тільки адміну).",
        "operationId": "get_user_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Оновити користувача",
        "description": "Оновлює профіль користувача (тільки свій або адміну).",
        "operationId": "update_user_endpoint_users__user_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Видалити користувача (soft-delete)",
        "description": "М'яко видаляє користувача (тільки себе або адміну).",
        "operationId": "delete_user_endpoint_users__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/words": {
      "post": {
        "tags": [
          "Words"
        ],
        "summary": "Додати слово у словник",
        "description": "Додає нове слово до загального словника.",
        "operationId": "create_word_words_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WordCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WordRead"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Words"
        ],
        "summary": "Отримати всі слова із словника",
        "description": "Повертає всі не видалені слова (можна фільтрувати через skip/limit).",
        "operationId": "list_words_words_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Пропустити N записів",
              "default": 0,
              "title": "Skip"
            },
            "description": "Пропустити N записів"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Скільки записів повернути",
              "default": 100,
              "title": "Limit"
            },
            "description": "Скільки записів повернути"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WordRead"
                  },
                  "title": "Response List Words Words Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/words/{word_id}": {
      "get": {
        "tags": [
          "Words"
        ],
        "summary": "Отримати слово за id",
        "description": "Повертає слово по id.",
        "operationId": "get_word_words__word_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "word_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Word Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WordRead"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Words"
        ],
        "summary": "Оновити слово у словнику",
        "description": "Оновлює слово у словнику (тільки адмін).",
        "operationId": "update_word_words__word_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "word_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Word Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WordUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WordRead"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Words"
        ],
        "summary": "Видалити слово зі словника (soft-delete)",
        "description": "М'яко видаляє слово зі словника (тільки адмін).",
        "operationId": "delete_word_words__word_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "word_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Word Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserRead": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Унікальний email користувача"
          },
          "username": {
            "type": "string",
            "maxLength": 150,
            "minLength": 2,
            "title": "Username",
            "description": "Унікальне ім'я користувача (логін)"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Унікальний ідентифікатор користувача"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Чи активний користувач"
          },
          "is_superuser": {
            "type": "boolean",
            "title": "Is Superuser",
            "description": "Чи є користувач адміністратором"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At",
            "description": "Дата реєстрації"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "Дата останнього оновлення"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At",
            "description": "Дата видалення користувача (якщо є)"
          }
        },
        "type": "object",
        "required": [
          "email",
          "username",
          "id",
          "is_active",
          "is_superuser"
        ],
        "title": "UserRead",
        "description": "Схема для читання (API-відповіді) користувача."
      },
...
      "UserUpdate": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "Оновлений email"
          },
          "username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Username",
            "description": "Оновлений логін"
          },
          "password": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Password",
            "description": "Новий пароль"
          }
        },
        "type": "object",
        "title": "UserUpdate",
        "description": "Схема для оновлення даних користувача."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WordCreate": {
        "properties": {
          "base_form": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Base Form",
            "description": "Початкова форма слова (без артикля)"
          },
          "lang_from": {
            "type": "string",
            "maxLength": 10,
            "minLength": 2,
            "title": "Lang From",
            "description": "Мова оригіналу (ISO-код, напр. 'de', 'en')"
          },
          "lang_to": {
            "type": "string",
            "maxLength": 10,
            "minLength": 2,
            "title": "Lang To",
            "description": "Мова перекладу (ISO-код, напр. 'uk', 'en')"
          },
          "article": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Article",
            "description": "Артикль (напр., 'der', 'die', 'das') — генерується GPT"
          },
          "base_and_article": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Base And Article",
            "description": "Форма слова з артиклем — генерується GPT"
          },
          "translation": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation",
            "description": "Переклад слова — генерується GPT"
          },
          "example": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Example",
            "description": "Приклад вживання слова — генерується GPT"
          },
          "level": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Level",
            "description": "CEFR рівень слова (A1, A2, B1, B2, C1, C2) — генерується GPT"
          }
        },
        "type": "object",
        "required": [
          "base_form",
          "lang_from",
          "lang_to"
        ],
        "title": "WordCreate",
        "description": "Схема для створення слова у словнику."
      },
      "WordRead": {
        "properties": {
          "base_form": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Base Form",
            "description": "Початкова форма слова (без артикля)"
          },
          "lang_from": {
            "type": "string",
            "maxLength": 10,
            "minLength": 2,
            "title": "Lang From",
            "description": "Мова оригіналу (ISO-код, напр. 'de', 'en')"
          },
          "lang_to": {
            "type": "string",
            "maxLength": 10,
            "minLength": 2,
            "title": "Lang To",
            "description": "Мова перекладу (ISO-код, напр. 'uk', 'en')"
          },
          "article": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Article",
            "description": "Артикль (напр., 'der', 'die', 'das') — генерується GPT"
          },
          "base_and_article": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Base And Article",
            "description": "Форма слова з артиклем — генерується GPT"
          },
          "translation": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation",
            "description": "Переклад слова — генерується GPT"
          },
          "example": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Example",
            "description": "Приклад вживання слова — генерується GPT"
          },
          "level": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Level",
            "description": "CEFR рівень слова (A1, A2, B1, B2, C1, C2) — генерується GPT"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Унікальний ідентифікатор слова"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Дата створення слова"
          },
          "deleted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At",
            "description": "Дата видалення (якщо застосовано)"
          }
        },
        "type": "object",
        "required": [
          "base_form",
          "lang_from",
          "lang_to",
          "id",
          "created_at"
        ],
        "title": "WordRead",
        "description": "Схема для читання слова (API-відповідь)."
      },
      "WordUpdate": {
        "properties": {
          "base_form": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Form",
            "description": "Оновлена початкова форма слова"
          },
          "article": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Article",
            "description": "Оновлений артикль"
          },
          "base_and_article": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Base And Article",
            "description": "Оновлена форма слова з артиклем"
          },
          "translation": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation",
            "description": "Оновлений переклад"
          },
          "lang_from": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Lang From",
            "description": "Оновлена мова оригіналу"
          },
          "lang_to": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Lang To",
            "description": "Оновлена мова перекладу"
          },
          "example": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Example",
            "description": "Оновлений приклад вживання"
          },
          "level": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Level",
            "description": "Оновлений CEFR рівень слова"
          }
        },
        "type": "object",
        "title": "WordUpdate",
        "description": "Схема для часткового оновлення даних слова."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}